backend/requirements.txt
# Core Django
Django>=5.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# AI Services
openai>=1.0.0
groq>=0.4.0

# Image Processing
Pillow>=10.0.0
reportlab>=4.0.0

# Cloud Storage
boto3>=1.34.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
django-ratelimit>=4.1.0
cryptography>=41.0.0
argon2-cffi>=23.1.0
pyotp>=2.9.0

# OAuth & Multi-Provider Authentication
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
PyGithub>=2.1.0
firebase-admin>=6.4.0
python-jose>=3.3.0

# Security & Protection
django-csp>=3.7
bleach>=6.1.0
validators>=0.22.0


# Caching & Task Queue
redis>=5.0.0
django-redis>=5.4.0
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# API Documentation
drf-spectacular>=0.27.0
drf-nested-routers>=0.93.0

# Logging
python-json-logger>=2.0.0

# Monitoring & Debugging
django-debug-toolbar>=4.2.0
sentry-sdk>=1.40.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Payment Processing
stripe>=7.0.0

# Real-time & Webhooks
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Search
django-elasticsearch-dsl>=7.4.0
elasticsearch>=9.0.0,<10.0.0

# Export & Document Generation
reportlab>=4.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
